INCOMPLETE: Add layerReadAccess
Wrap the _entirety_ of the synchronization logic for readers
into a function that deals with locking, reloads, and
in-process locks.

Almost all readers should only access layerStore through
layerReadAccess (though there _might_ be some exceptions
where several stores need to be locked simultaneously, or
something like that).

Also introduce layerReadToken to make the locking assertions
visible in the type system, and much harder not to get right.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>