INCOMPLETE: Add layerWriteAccess
Wrap the _entirety_ of the synchronization logic for writers
into a function that deals with locking, reloads, and
in-process locks.

Almost all readers should only access layerStore through
layerWriteAccess (though there _might_ be some exceptions
where several stores need to be locked simultaneously, or
something like that).

Also introduce layerWriteToken to make the locking assertions
visible in the type system, and much harder not to get right;
and use it to remove a run-time assertion as one example.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>